ci: fix workflows (#10)
Fixes the following workflows:

- **Current with Charmcraft** - This workflow builds Charmcraft from
`main` and checks that `charmcraft init` generates the same charms as
stored in this repo.

The workflow was failing because `make setup` couldn't download
https://security.ubuntu.com/ubuntu/pool/universe/p/python-apt/python-apt-dev_2.7.7ubuntu5_all.deb.
After testing in a local VM, I found that I needed to run `sudo apt-get
update` before `make setup`. This PR makes this addition to the
workflow.

- **Charm checks** - This workflow includes a check that the
implementation script for kubernetes-extra produces the same charm as
stored in the repo. The check was failing because some transitive
dependencies of `requests` had had version bumps. For kubernetes-extra,
we only care about primary dependencies added to `pyproject.toml`, so
I've excluded `uv.lock` from the diff and specified an exact version of
`requests`.