ci: properly detect charm changes (#11)
Our CI was subtly broken. In a couple of places, we checked that
generated charms matched what's stored in the repo. However, we used
`git diff`, which doesn't account for untracked files. If a generated
charm had introduced a new file, we wouldn't have caught it.

This PR switches to a more robust method based on `git status`, with a
dedicated Just target: `just diff`.